+++ src/helpers/check-mobile.js
