+++ src/helpers/check-mobile.test.js
